Bayesian Test from Outlier Scores

Test for no adverse shift with outlier scores. Like goodness-of-fit testing, this two-sample comparison takes the training (outlier) scores, os_train, as the reference. The method checks whether the test scores, os_test, are worse off relative to the training set.

bf_from_os(os_train, os_test, n_pt = 4000, adverse_threshold = 1/12)

Arguments

os_train

Outlier scores in training (reference) set.

os_test

Outlier scores in test set.

n_pt

The number of permutations.

adverse_threshold

WAUC threshold for adverse shift.

Value

A named list of class outlier.bayes containing:

posterior: Posterior distribution of WAUC test statistic

adverse_threshold: WAUC threshold for adverse shift

adverse_probability: probability of adverse shift

bayes_factor: Bayes factor

outlier_scores: outlier scores from training and test set

Details

The posterior distribution of the WAUC test statistic is based on n_pt (boostrap) permutations. The method relies on the Bayesian bootstrap as a resampling procedure to smooth the instance weights as in Gu et al (2008). This framework borrows from Johnson (2005), which shows how to turn a test statistic, which may very well be frequentist, into its Bayesian counterpart, the bayes factor.

References

Kamulete, V. M. (2023). Are you OK? A Bayesian sequential test for adverse shift. Manuscript in preparation.

Johnson, V. E. (2005). Bayes factors based on test statistics. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 67(5), 689-701.

Gu, J., Ghosal, S., & Roy, A. (2008). Bayesian bootstrap estimation of ROC curve. Statistics in medicine, 27(26), 5407-5420.

See also

Examples

# \donttest{
library(dsos)
set.seed(12345)
os_train <- rnorm(n = 100)
os_test <- rnorm(n = 100)
bayes_test <- bf_from_os(os_train, os_test)
bayes_test
#> 	Bayesian test for no adverse shift 
#> 
#> Bayes factor (BF) = 0.07, cutoff (weighted AUC/WAUC) = 0.0833
#> 
#> Model: bayesian bootstrap with 4000 replicates (simulations) 
#> BF's numerator: Pr(WAUC >= 0.0833) 
#> BF's denominator: Pr(WAUC < 0.0833) 
#> => BF > 3 favors view that the test set is worse off than training.
#> Sample sizes: 100 in training and 100 in test set.
# Run in parallel on local cluster
library(future)
future::plan(future::multisession)
parallel_test <- bf_from_os(os_train, os_test)
parallel_test
#> 	Bayesian test for no adverse shift 
#> 
#> Bayes factor (BF) = 0.06, cutoff (weighted AUC/WAUC) = 0.0833
#> 
#> Model: bayesian bootstrap with 4000 replicates (simulations) 
#> BF's numerator: Pr(WAUC >= 0.0833) 
#> BF's denominator: Pr(WAUC < 0.0833) 
#> => BF > 3 favors view that the test set is worse off than training.
#> Sample sizes: 100 in training and 100 in test set.
# }