Permutation Test from Outlier Scores

Test for no adverse shift with outlier scores. Like goodness-of-fit testing, this two-sample comparison takes the training set, x_train or os_train, as the reference. The method checks whether the test set, x_test or os_test, is worse off relative to this reference set.

pt_from_os(os_train, os_test, n_pt = 2000)

Arguments

os_train

Outlier scores in training (reference) set.

os_test

Outlier scores in test set.

n_pt

The number of permutations.

Value

A named list of class outlier.test containing:

statistic: observed WAUC statistic

seq_mct: sequential Monte Carlo test, when applicable

p_value: p-value

outlier_scores: outlier scores from training and test set

Details

The empirical null distribution uses n_pt permutations to estimate the p-value. For speed, this is implemented as a sequential Monte Carlo test with the simctest package. See Gandy (2009) for details. The prefix pt refers to permutation test. This approach does not use the asymptotic null distribution for the weighted AUC (WAUC), the test statistic. This is the recommended approach for small samples.

Notes

These outlier scores should all be out-of-bag scores to mimic out-of-sample behaviour. Otherwise, the scores from the training (reference) distribution are biased (overfitted) whereas those from the test set are not. This mismatch -- in-sample training scores versus out-of-sample (out-of-bag) test scores -- would void the validity of the statistical test. A simple fix for this, without using resampling and/or permutations, is to get the training (reference) scores from a fresh (unused) validation set.

References

Kamulete, V. M. (2022). Test for non-negligible adverse shifts. In The 38th Conference on Uncertainty in Artificial Intelligence. PMLR.

Gandy, A. (2009). Sequential implementation of Monte Carlo tests with uniformly bounded resampling risk. Journal of the American Statistical Association, 104(488), 1504-1511.

Li, J., & Fine, J. P. (2010). Weighted area under the receiver operating characteristic curve and its application to gene selection. Journal of the Royal Statistical Society: Series C (Applied Statistics), 59(4), 673-692.

See also

Examples

# \donttest{
library(dsos)
set.seed(12345)
os_train <- rnorm(n = 100)
os_test <- rnorm(n = 100)
test_result <- pt_from_os(os_train, os_test)
test_result
#> 
#> 	Test for no adverse shift
#> 
#> p-value = 1
#> 
#> The first 5 outlier scores (out of 100) from the training (reference) sample:
#> [1]  0.585529  0.709466 -0.109303 -0.453497  0.605887
#> 
#> The first 5 outlier scores (out of 100) from the test sample:
#> [1]  0.223925 -1.156223  0.422419 -1.324755  0.141084
#> 
# }