Permutation Test With Out-Of-Bag Scores

Test for no adverse shift with outlier scores. Like goodness-of-fit testing, this two-sample comparison takes the training set, x_train or os_train, as the reference. The method checks whether the test set, x_test or os_test, is worse off relative to this reference set.

pt_oob(x_train, x_test, scorer, n_pt = 2000)

Arguments

x_train

Training (reference/validation) sample.

x_test

Test sample.

scorer

Function which returns a named list with outlier scores from the training and test sample. The first argument to scorer must be x_train; the second, x_test. The returned named list contains two elements: train and test, each of which is a vector of (outlier) scores. See notes below for more information.

n_pt

The number of permutations.

Value

A named list of class outlier.test containing:

statistic: observed WAUC statistic

seq_mct: sequential Monte Carlo test, when applicable

p_value: p-value

outlier_scores: outlier scores from training and test set

Details

The empirical null distribution uses n_pt permutations to estimate the p-value. For speed, this is implemented as a sequential Monte Carlo test with the simctest package. See Gandy (2009) for details. The prefix pt refers to permutation test. This approach does not use the asymptotic null distribution for the weighted AUC (WAUC), the test statistic. This is the recommended approach for small samples.

Notes

The scoring function, scorer, predicts out-of-bag scores to mimic out-of-sample behaviour. The suffix oob stands for out-of-bag to highlight this point. This out-of-bag variant avoids refitting the underlying algorithm from scorer at every permutation. It can, as a result, be computionally appealing.

References

Kamulete, V. M. (2022). Test for non-negligible adverse shifts. In The 38th Conference on Uncertainty in Artificial Intelligence. PMLR.

Gandy, A. (2009). Sequential implementation of Monte Carlo tests with uniformly bounded resampling risk. Journal of the American Statistical Association, 104(488), 1504-1511.

Li, J., & Fine, J. P. (2010). Weighted area under the receiver operating characteristic curve and its application to gene selection. Journal of the Royal Statistical Society: Series C (Applied Statistics), 59(4), 673-692.

See also

Examples

# \donttest{
library(dsos)
set.seed(12345)
data(iris)
idx <- sample(nrow(iris), 2 / 3 * nrow(iris))
xy_train <- iris[idx, ]
xy_test <- iris[-idx, ]

# First example: residual diagnostics
scorer_1 <- function(x_train, x_test) score_rd(x_train, x_test, response_name = "Species")
rd_test <- pt_oob(xy_train, xy_test, scorer = scorer_1)
rd_test
#> 
#> 	Test for no adverse shift
#> 
#> p-value = 0.034
#> 
#> The first 5 outlier scores (out of 100) from the training (reference) sample:
#> [1] 0.00865036 0.00000000 0.00000000 0.00000000 0.00000000
#> 
#> The first 5 outlier scores (out of 50) from the test sample:
#> [1] 0 0 0 0 0
#> 

# Second example: prediction uncertainty
scorer_2 <- function(x_train, x_test) score_rue(x_train, x_test, response_name = "Species")
rue_test <- pt_oob(xy_train, xy_test, scorer = scorer_2)
rue_test
#> 
#> 	Test for no adverse shift
#> 
#> p-value = 0.141
#> 
#> The first 5 outlier scores (out of 100) from the training (reference) sample:
#> [1] 0.00561083 0.00588218 0.00000000 0.00000000 0.00000000
#> 
#> The first 5 outlier scores (out of 50) from the test sample:
#> [1] 0 0 0 0 0
#> 

# Third example: sample memberships (class probabilities)
setosa <- iris[1:50, 1:4] # Training sample: Species == 'setosa'
versicolor <- iris[51:100, 1:4] # Test sample: Species == 'versicolor'
scorer_3 <- function(x_train, x_test) score_cp(x_train, x_test)
cp_test <- pt_oob(setosa, versicolor, scorer = scorer_3)
cp_test
#> 
#> 	Test for no adverse shift
#> 
#> p-value < 0.00000000000000022
#> 
#> The first 5 outlier scores (out of 50) from the training (reference) sample:
#> [1] 0 0 0 0 0
#> 
#> The first 5 outlier scores (out of 50) from the test sample:
#> [1] 1 1 1 1 1
#> 
# }