Predict Isolation Scores

Predict isolation scores using (extended) isolation forest with the isotree package. The prefix od stands for outlier detection, the relevant notion of outlyingness. This function is useful to test for dataset shift via density-based scores: isolation scores are inversely related, if not quite proportional, to densities.

score_od(x_train, x_test, n_trees = 500L, threshold = 0.6)

Arguments

x_train

Training (reference) sample.

x_test

Test sample.

n_trees

The number of trees in isolation forest.

threshold

Decision threshold. Set to a default value of 0.6, following Chabchoub et al. (2022). Outlier scores higher than the threshold are considered outliers, and lower values are inliers.

Value

A named list or object of class outlier.test containing:

train: vector of scores in training set

test: vector of scores in test set

Notes

Isolation forest detects isolated points that are typically out-of-distribution relative to the high-density regions of the data distribution. Any performant method for density-based out-of-distribution detection can replace isolation forest. The decision threshold, threshold, clips (winsorizes) the scores so that lower scores are set to the threshold value.

References

Liu, F. T., Ting, K. M., & Zhou, Z. H. (2008, December). Isolation forest. In 2008 Eighth IEEE International Conference on Data Mining (pp. 413-422). IEEE.

Chabchoub, Y., Togbe, M. U., Boly, A., & Chiky, R. (2022). An in-depth study and improvement of Isolation Forest.. IEEE Access, 10, 10219-10237.

#' @details score_od first fits to the training data and then predict in-sample for this reference sample. Then it predicts out-of-sample for the test set. As a result, estimating p-value via permutations require refitting the algorithm for every permutation.

See also

Examples

# \donttest{
library(dsos)
set.seed(12345)
data(iris)
setosa <- iris[1:50, 1:4] # Training sample: Species == 'setosa'
versicolor <- iris[51:100, 1:4] # Test sample: Species == 'versicolor'
score_od(setosa, versicolor)
#> $test
#>       51       52       53       54       55       56       57       58 
#> 0.674343 0.672536 0.674081 0.673545 0.672993 0.672098 0.676666 0.664948 
#>       59       60       61       62       63       64       65       66 
#> 0.673706 0.673943 0.668086 0.673942 0.665721 0.674062 0.670799 0.673192 
#>       67       68       69       70       71       72       73       74 
#> 0.675250 0.666555 0.675013 0.667670 0.678094 0.672884 0.674463 0.671084 
#>       75       76       77       78       79       80       81       82 
#> 0.672640 0.673440 0.673128 0.675084 0.673138 0.665407 0.669511 0.665245 
#>       83       84       85       86       87       88       89       90 
#> 0.670312 0.676072 0.676918 0.678196 0.674158 0.670614 0.673113 0.673591 
#>       91       92       93       94       95       96       97       98 
#> 0.671322 0.672497 0.670786 0.668197 0.672671 0.672231 0.673275 0.671262 
#>       99      100 
#> 0.666917 0.671606 
#> 
#> $train
#>        1        2        3        4        5        6        7        8 
#> 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 
#>        9       10       11       12       13       14       15       16 
#> 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 
#>       17       18       19       20       21       22       23       24 
#> 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 
#>       25       26       27       28       29       30       31       32 
#> 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 
#>       33       34       35       36       37       38       39       40 
#> 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 
#>       41       42       43       44       45       46       47       48 
#> 0.600000 0.604261 0.600000 0.600000 0.600000 0.600000 0.600000 0.600000 
#>       49       50 
#> 0.600000 0.600000 
#> 
# }