Predict Out-of-bag Errors (Residuals)

Predict out-of-bag errors (residuals) using random forest with the ranger package. The prefix rd stands for residual diagnostic, the relevant notion of outlyingness. This function is useful to test for dataset shift via prediction errors from the underlying supervised algorithm.

score_rd(x_train, x_test, n_trees = 500L, response_name = "label")

Arguments

x_train

Training (reference) sample.

x_test

Test sample.

n_trees

The number of trees in random forest.

response_name

The column name of the categorical outcome to predict.

Value

A named list or object of class outlier.test containing:

train: vector of scores in training set

test: vector of scores in test set

Details

score_rd first fits to the training data and uses out-of-bag predictions to estimate errors (residuals) for this reference sample. Then it leverages out-of-sample predictions to calculate errors for the test set. As a result, estimating p-value via permutations does not require refitting the algorithm for every permutation. See Hediger et al. (2022) for details on this approach.

Notes

Residuals traditionally underpin diagnostics (misspecification) tests in supervised learning. For a contemporaneous example of this approach using machine learning, see Janková et al. (2020) and references therein.

References

Janková, J., Shah, R. D., Bühlmann, P., & Samworth, R. J. (2020). Goodness-of-fit testing in high dimensional generalized linear models. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 82(3), 773-795.

Hediger, S., Michel, L., & Näf, J. (2022). On the use of random forest for two-sample testing. Computational Statistics & Data Analysis, 170, 107435.

See also

Examples

# \donttest{
library(dsos)
set.seed(12345)
data(iris)
idx <- sample(nrow(iris), 2 / 3 * nrow(iris))
xy_train <- iris[idx, ]
xy_test <- iris[-idx, ]
score_rd(xy_train, xy_test, n_trees = 500L, response_name = "Species")
#> $test
#>  [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [7] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.18645714 0.00000000
#> [19] 0.33800000 0.33800000 0.00000000 0.00000000 0.33800000 0.90304286
#> [25] 0.13202857 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [31] 0.00000000 0.01218810 0.98800000 0.00867143 0.00000000 0.02405952
#> [37] 0.00974048 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [43] 0.09513810 0.00000000 0.75750000 0.76617143 0.76391429 0.00000000
#> [49] 0.00000000 0.00000000
#> 
#> $train
#>   [1] 0.00865036 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>   [7] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02139037
#>  [13] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [19] 0.00000000 0.00000000 0.00000000 0.00000000 0.02155844 0.00000000
#>  [25] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [31] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.01339114
#>  [37] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [43] 0.01961580 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [49] 0.00000000 0.95265761 0.00000000 0.00000000 0.00000000 0.00000000
#>  [55] 0.00000000 0.00000000 0.00000000 0.01046919 0.00000000 0.00000000
#>  [61] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [67] 0.00000000 0.00120482 0.00000000 0.00000000 0.00000000 0.00000000
#>  [73] 0.00000000 0.03554320 0.00000000 0.00000000 0.02424242 0.00000000
#>  [79] 0.32776455 0.02964559 0.00000000 0.00000000 0.00000000 0.44949495
#>  [85] 0.00000000 0.00000000 0.00000000 0.03025847 0.00000000 0.02649007
#>  [91] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.39333333
#>  [97] 0.02902560 0.00000000 0.25301205 0.00000000
#> 
# }