Predict Resampling Uncertainty (Prediction Confidence)

Estimate prediction uncertainty using random forest with the ranger package. The prefix rue stands for resampling uncertainty estimation, the relevant notion of outlyingness. This function is useful to test for dataset shift via prediction uncertainty from supervised algorithms.

score_rue(x_train, x_test, n_trees = 500L, response_name = "label")

Arguments

x_train

Training (reference) sample.

x_test

Test sample.

n_trees

The number of trees in random forest.

response_name

The column name of the categorical outcome to predict.

Value

A named list or object of class outlier.test containing:

train: vector of scores in training set

test: vector of scores in test set

Details

score_rue first fits to the training data and uses out-of-bag predictions to estimate prediction uncertainty for this reference sample. Then it leverages out-of-sample predictions to do the same for the test set. As a result, estimating p-value via permutations does not require refitting the algorithm for every permutation.

Notes

For prediction uncertainty, we essentially implement the approach in Schulam & Saria (2019) with random forest. The standard errors of the mean predictions are the outlier scores. Any performant method for confidence-based out-of-distribution (OOD) detection can replace random forest. Berger et al. (2021) compares methods for confidence-based OOD detection.

References

Schulam, P., & Saria, S. (2019, April). Can you trust this prediction? Auditing pointwise reliability after learning. In The 22nd International Conference on Artificial Intelligence and Statistics (pp. 1022-1031). PMLR.

Berger, C., Paschali, M., Glocker, B., & Kamnitsas, K. (2021). Confidence-based out-of-distribution detection: a comparative study and analysis. In Uncertainty for Safe Utilization of Machine Learning in Medical Imaging, and Perinatal Imaging, Placental and Preterm Image Analysis (pp. 122-132). Springer, Cham.

See also

Examples

# \donttest{
library(dsos)
set.seed(12345)
data(iris)
idx <- sample(nrow(iris), 2 / 3 * nrow(iris))
xy_train <- iris[idx, ]
xy_test <- iris[-idx, ]
outlier_scores <- score_rue(xy_train, xy_test, response_name = "Species")
str(outlier_scores)
#> List of 2
#>  $ test : num [1:50] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ train: num [1:100] 0.00386 0 0 0 0 ...
# }